CRUD notes works for different users

+++ Notes Application/DataLayer/NoteRepository.cs
@@ -36,15 +36,16 @@ namespace DataLayer
             //constr = "Data Source=DESKTOP-PCL70MC\\SQLEXPRESS;Initial Catalog=test;Integrated Security=True";
         }
 
-        public void CreateNote(string title, string text)
+        public void CreateNote(int userId, string title, string text)
         {
             conn = new SqlConnection(constr);
             conn.Open();
             SqlCommand cmd;
 
-            string sql = "INSERT INTO notes VALUES(1, @title, @text, @date, @editdate)";
+            string sql = "INSERT INTO notes VALUES(@userid, @title, @text, @date, @editdate)";
 
             cmd = new SqlCommand(sql, conn);
+            cmd.Parameters.Add(new SqlParameter { ParameterName = "@userid", Value = userId });
             cmd.Parameters.Add(new SqlParameter { ParameterName = "@title", Value = title });
             cmd.Parameters.Add(new SqlParameter { ParameterName = "@text", Value = text });
             cmd.Parameters.Add(new SqlParameter { ParameterName = "@date", Value = DateTime.Now });
@@ -65,16 +66,17 @@ namespace DataLayer
             }
         }
 
-        public List<NoteDTO> ReadNotes()
+        public List<NoteDTO> ReadNotes(int userId)
         {
             conn = new SqlConnection(constr);
             conn.Open();
             SqlCommand cmd;
             SqlDataReader dreader;
 
-            string sql = "SELECT Id,Title,[Text] FROM notes";
+            string sql = "SELECT Id,Title,[Text] FROM notes WHERE UserId=@userid";
 
             cmd = new SqlCommand(sql, conn);
+            cmd.Parameters.Add(new SqlParameter { ParameterName = "@userid", Value = userId });
 
             List<NoteDTO> notesDTO = new List<NoteDTO>();
 
@@ -108,16 +110,17 @@ namespace DataLayer
             return notesDTO;
         }
 
-        public void UpdateNote( int id, string title, string text)
+        public void UpdateNote( int noteId, int userId, string title, string text)
         {
             conn = new SqlConnection(constr);
             conn.Open();
             SqlCommand cmd;
 
-            string sql = "UPDATE notes set Title=@title, [Text]=@text, EditDate=@editdate WHERE Id= @id";
+            string sql = "UPDATE notes set Title=@title, [Text]=@text, EditDate=@editdate WHERE Id= @noteid AND UserId=@userid";
 
             cmd = new SqlCommand(sql, conn);
-            cmd.Parameters.Add(new SqlParameter { ParameterName = "@id", Value = id });
+            cmd.Parameters.Add(new SqlParameter { ParameterName = "@noteid", Value = noteId });
+            cmd.Parameters.Add(new SqlParameter { ParameterName = "@userid", Value = userId });
             cmd.Parameters.Add(new SqlParameter { ParameterName = "@title", Value = title });
             cmd.Parameters.Add(new SqlParameter { ParameterName = "@text", Value = text });
             cmd.Parameters.Add(new SqlParameter { ParameterName = "@editdate", Value = DateTime.Now });
@@ -136,16 +139,17 @@ namespace DataLayer
                 conn.Close();
             }
         }
-        public void DeleteNote(int id)
+        public void DeleteNote(int noteId, int userId)
         {
             conn = new SqlConnection(constr);
             conn.Open();
             SqlCommand cmd;
 
-            string sql = "DELETE FROM notes WHERE Id=@id";
+            string sql = "DELETE FROM notes WHERE Id=@id AND UserId=@userid";
 
             cmd = new SqlCommand(sql, conn);
-            cmd.Parameters.Add(new SqlParameter { ParameterName = "@id", Value = id });
+            cmd.Parameters.Add(new SqlParameter { ParameterName = "@id", Value = noteId });
+            cmd.Parameters.Add(new SqlParameter { ParameterName = "@userid", Value = userId });
 
             try
             {
